+++ TrainingApp/src/App.js
@@ -14,7 +14,6 @@ import navigation from './navigation';
 const App: () => React$Node = () => {
   return (
     <>
-      {/* <SignUp></SignUp> */}
       <MainNavigator ref={navigation.init} />
     </>
   );
