+++ TrainingApp/src/constants/endPoints.js
@@ -1,0 +1,3 @@
+export const AUTH_URL = 'https://identitytoolkit.googleapis.com/v1';
+
+export const BASE_URL = 'https://rn-complete-guide.firebaseio.com';
