+++ TrainingApp/src/models/order.js
@@ -9,13 +9,6 @@ class Order {
   }
 
   get readableDate() {
-    //   return this.date.toLocaleDateString('en-EN', {
-    //       year: 'numeric',
-    //       month: 'long',
-    //       day: 'numeric',
-    //       hour: '2-digit',
-    //       minute: '2-digit'
-    //   });
     return moment(this.date).format('MMMM Do YYYY, hh:mm');
   }
 }
