+++ TrainingApp/src/store/actions/auth.js
@@ -1,9 +1,6 @@
 import { AsyncStorage } from 'react-native';
-
-// export const SIGNUP = 'SIGNUP';
-// export const LOGIN = 'LOGIN';
-export const AUTHENTICATE = 'AUTHENTICATE';
-export const LOGOUT = 'LOGOUT';
+import { AUTHENTICATE, LOGOUT } from './actionTypes';
+import { AUTH_URL } from '../../constants/endPoints';
 
 let timer;
 
@@ -17,7 +14,7 @@ export const authenticate = (userId, token, expiryTime) => {
 export const signup = (email, password) => {
   return async dispatch => {
     const response = await fetch(
-      'https://identitytoolkit.googleapis.com/v1/accounts:signUp?key=AIzaSyBY8UJq_xLD0nEe1HZHuvEOUfYIS9gg4pA',
+      `${AUTH_URL}/accounts:signUp?key=AIzaSyBY8UJq_xLD0nEe1HZHuvEOUfYIS9gg4pA`,
       {
         method: 'POST',
         headers: {
@@ -60,7 +57,7 @@ export const signup = (email, password) => {
 export const login = (email, password) => {
   return async dispatch => {
     const response = await fetch(
-      'https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=AIzaSyBY8UJq_xLD0nEe1HZHuvEOUfYIS9gg4pA',
+      `${AUTH_URL}/accounts:signInWithPassword?key=AIzaSyBY8UJq_xLD0nEe1HZHuvEOUfYIS9gg4pA`,
       {
         method: 'POST',
         headers: {
