+++ TrainingApp/src/store/actions/cart.js
@@ -1,5 +1,4 @@
-export const ADD_TO_CART = 'ADD_TO_CART';
-export const REMOVE_FROM_CART = 'REMOVE_FROM_CART';
+import { ADD_TO_CART, REMOVE_FROM_CART } from './actionTypes';
 
 export const addToCart = product => {
   return { type: ADD_TO_CART, product: product };
