+++ TrainingApp/src/store/actions/products.js
@@ -1,9 +1,11 @@
 import Product from '../../models/product';
-
-export const DELETE_PRODUCT = 'DELETE_PRODUCT';
-export const CREATE_PRODUCT = 'CREATE_PRODUCT';
-export const UPDATE_PRODUCT = 'UPDATE_PRODUCT';
-export const SET_PRODUCTS = 'SET_PRODUCTS';
+import {
+  DELETE_PRODUCT,
+  CREATE_PRODUCT,
+  UPDATE_PRODUCT,
+  SET_PRODUCTS
+} from './actionTypes';
+import { BASE_URL } from '../../constants/endPoints';
 
 export const fetchProducts = () => {
   return async (dispatch, getState) => {
@@ -11,7 +13,7 @@ export const fetchProducts = () => {
     const userId = getState().auth.userId;
     try {
       const response = await fetch(
-        'https://rn-complete-guide.firebaseio.com/products.json'
+        '${BASE_URL}/products.json'
       );
 
       if (!response.ok) {
@@ -50,7 +52,7 @@ export const deleteProduct = productId => {
   return async (dispatch, getState) => {
     const token = getState().auth.token;
     const response = await fetch(
-      `https://rn-complete-guide.firebaseio.com/products/${productId}.json?auth=${token}`,
+      `${BASE_URL}/products/${productId}.json?auth=${token}`,
       {
         method: 'DELETE'
       }
@@ -69,7 +71,7 @@ export const createProduct = (title, description, imageUrl, price) => {
     const token = getState().auth.token;
     const userId = getState().auth.userId;
     const response = await fetch(
-      `https://rn-complete-guide.firebaseio.com/products.json?auth=${token}`,
+      `${BASE_URL}/products.json?auth=${token}`,
       {
         method: 'POST',
         headers: {
@@ -105,7 +107,7 @@ export const updateProduct = (id, title, description, imageUrl) => {
   return async (dispatch, getState) => {
     const token = getState().auth.token;
     const response = await fetch(
-      `https://rn-complete-guide.firebaseio.com/products/${id}.json?auth=${token}`,
+      `${BASE_URL}/products/${id}.json?auth=${token}`,
       {
         method: 'PATCH',
         headers: {
