+++ TrainingApp/src/store/reducers/cart.js
@@ -1,7 +1,5 @@
-import { ADD_TO_CART, REMOVE_FROM_CART } from '../actions/cart';
-import { ADD_ORDER } from '../actions/orders';
+import { ADD_TO_CART, REMOVE_FROM_CART, ADD_ORDER, DELETE_PRODUCT } from '../actions/actionTypes';
 import CartItem from '../../models/cart-item';
-import { DELETE_PRODUCT } from '../actions/products';
 
 const initialState = {
   items: {},
